Move cstring and ccalback to abi_utils module
