+++ libvcx/src/abi_utils/cstring.rs
