+++ libvcx/src/api/credential.rs
@@ -6,7 +6,7 @@ use libc::c_char;
 use crate::connection;
 use crate::credential;
 use crate::error::prelude::*;
-use crate::utils::cstring::CStringUtils;
+use crate::abi_utils::cstring::CStringUtils;
 use crate::utils::error;
 use crate::utils::runtime::execute;
