+++ libvcx/src/api/disclosed_proof.rs
@@ -5,7 +5,7 @@ use libc::c_char;
 
 use crate::{connection, disclosed_proof};
 use crate::error::prelude::*;
-use crate::utils::cstring::CStringUtils;
+use crate::abi_utils::cstring::CStringUtils;
 use crate::utils::error;
 use crate::utils::runtime::execute;
