+++ libvcx/src/api/utils.rs
@@ -12,7 +12,7 @@ use crate::connection;
 use crate::error::prelude::*;
 use crate::libindy::utils::payments;
 use crate::utils::constants::*;
-use crate::utils::cstring::CStringUtils;
+use crate::abi_utils::cstring::CStringUtils;
 use crate::utils::error;
 use crate::utils::runtime::execute;
 use crate::utils::provision::AgentProvisionConfig;
